+++ src/Maestro/API/V1/Addresses.hs
@@ -39,6 +39,16 @@ data AddressesAPI route = AddressesAPI
       :> Pagination
       :> Get '[JSON] PaginatedOutputReferenceObject
 
+  , addressTxs
+      :: route
+      :- Capture "address" (Bech32StringOf Address)
+      :> "transactions"
+      :> QueryParam "order" Order
+      :> QueryParam "from" SlotNo
+      :> QueryParam "to" SlotNo
+      :> Pagination
+      :> Get '[JSON] PaginatedAddressTransaction
+
   , paymentCredentialUtxos
       :: route
       :- "cred"
@@ -46,8 +56,30 @@ data AddressesAPI route = AddressesAPI
       :> "utxos"
       :> QueryParam "resolve_datums" Bool
       :> QueryParam "with_cbor" Bool
+      :> QueryParam "asset" NonAdaNativeToken
       -- TODO: Support for more query parameters.
       :> Pagination
       :> Get '[JSON] PaginatedUtxoWithSlot
+
+  , paymentCredentialsUtxos
+      :: route
+      :- "cred"
+      :> "utxos"
+      :> QueryParam "resolve_datums" Bool
+      :> QueryParam "with_cbor" Bool
+      :> Pagination
+      :> ReqBody '[JSON] [Bech32StringOf PaymentCredentialAddress]
+      :> Post '[JSON] PaginatedUtxoWithSlot
+
+  , paymentCredentialTxs
+      :: route
+      :- "cred"
+      :> Capture "credential" (Bech32StringOf PaymentCredentialAddress)
+      :> "transactions"
+      :> QueryParam "order" Order
+      :> QueryParam "from" SlotNo
+      :> QueryParam "to" SlotNo
+      :> Pagination
+      :> Get '[JSON] PaginatedPaymentCredentialTransaction
 
   } deriving (Generic)