+++ src/Maestro/Client/V1.hs
@@ -3,6 +3,7 @@ module Maestro.Client.V1
   , module Maestro.Client.Error
   , module Maestro.Client.V1.Core
   , module Maestro.Client.V1.Addresses
+  , module Maestro.Client.V1.Assets
   , module Maestro.Client.V1.Blocks
   , module Maestro.Client.V1.Datum
   , module Maestro.Client.V1.DefiMarkets
@@ -16,6 +17,7 @@ module Maestro.Client.V1
 import           Maestro.Client.Env
 import           Maestro.Client.Error
 import           Maestro.Client.V1.Addresses
+import           Maestro.Client.V1.Assets
 import           Maestro.Client.V1.Blocks
 import           Maestro.Client.V1.Core
 import           Maestro.Client.V1.Datum